configuration: allow disable sync for specific folders

+++ folder-rsync/run.sh
@@ -7,33 +7,57 @@ CONFIG_PATH=/data/options.json
 rsyncserver=$(jq --raw-output ".rsyncserver" $CONFIG_PATH)
 rootfolder=$(jq --raw-output ".rootfolder" $CONFIG_PATH)
 username=$(jq --raw-output ".username" $CONFIG_PATH)
-password=$(jq --raw-output ".password" $CONFIG_PATH)
+syncconfig=$(jq --raw-output ".syncconfig" $CONFIG_PATH)
+syncaddons=$(jq --raw-output ".syncaddons" $CONFIG_PATH)
+syncbackup=$(jq --raw-output ".syncbackup" $CONFIG_PATH)
+syncshare=$(jq --raw-output ".syncshare" $CONFIG_PATH)
+syncssl=$(jq --raw-output ".syncssl" $CONFIG_PATH)
+syncmedia=$(jq --raw-output ".syncmedia" $CONFIG_PATH)
 
 rsyncurl="$username@$rsyncserver::$rootfolder"
 
 echo "[Info] trying to rsync hassio folders to $rsyncurl"
 echo ""
-echo "[Info] sync /config"
-sshpass -p $password rsync --no-perms -rltvh --delete --exclude '*.db-shm' --exclude '*.db-wal' /config/ $rsyncurl/config/ 
-echo ""
-echo "[Info] sync /addons"
-sshpass -p $password rsync --no-perms -rltvh --delete /addons/ $rsyncurl/addons/ 
-echo ""
-echo "[Info] sync /backup"
-sshpass -p $password rsync --no-perms -rltvh --delete /backup/ $rsyncurl/backup/ 
-echo ""
-echo "[Info] sync /share"
-sshpass -p $password rsync --no-perms -rltvh --delete /share/ $rsyncurl/share/ 
-echo ""
-echo "[Info] sync /ssl"
-sshpass -p $password rsync --no-perms -rltvh --delete /ssl/ $rsyncurl/ssl/ 
-if [ -d "/media" ]; then
- echo ""
- echo "[Info] sync /media"
- sshpass -p $password rsync --no-perms -rltvh --delete /media/ $rsyncurl/media/
-else 
- echo ""
- echo "[Info] /media not existing"
+
+if syncconfig then
+	echo "[Info] sync /config"
+	echo ""
+	sshpass -p $password rsync --no-perms -rltvh --delete --exclude '*.db-shm' --exclude '*.db-wal' /config/ $rsyncurl/config/ 
+fi
+
+if syncaddons then
+	echo "[Info] sync /addons"
+	echo ""
+	sshpass -p $password rsync --no-perms -rltvh --delete /addons/ $rsyncurl/addons/ 
+fi
+
+if syncbackup then
+	echo "[Info] sync /backup"
+	echo ""
+	sshpass -p $password rsync --no-perms -rltvh --delete /backup/ $rsyncurl/backup/ 
+fi
+
+if syncshare then
+	echo "[Info] sync /share"
+	echo ""
+	sshpass -p $password rsync --no-perms -rltvh --delete /share/ $rsyncurl/share/ 
+fi
+
+if syncssl then
+	echo "[Info] sync /ssl"
+	echo ""
+	sshpass -p $password rsync --no-perms -rltvh --delete /ssl/ $rsyncurl/ssl/ 
+fi
+
+if syncmedia then
+	if [ -d "/media" ]; then
+	 echo ""
+	 echo "[Info] sync /media"
+	 sshpass -p $password rsync --no-perms -rltvh --delete /media/ $rsyncurl/media/
+	else 
+	 echo ""
+	 echo "[Info] /media not existing"
+	fi
 fi
 
 echo "[Info] Finished rsync"
